+++ tests/testthat/test-seasonal_onset.R
@@ -148,3 +148,28 @@ test_that("Test that input argument checks work", {
   expect_error(seasonal_onset(tsd_data))
 
 })
+
+test_that("Test that selection of current and all seasons work as expected", {
+  start_date <- as.Date("2021-01-04")
+  end_date <- as.Date("2023-12-31")
+
+  weekly_dates <- seq.Date(from = start_date,
+                           to = end_date,
+                           by = "week")
+
+  obs <- stats::rpois(length(weekly_dates), 1000)
+
+  tsd_data <- to_time_series(
+    observation = obs,
+    time = as.Date(weekly_dates),
+    time_interval = "week"
+  )
+
+  current_season <- epi_calendar(end_date)
+
+  current_onset <- seasonal_onset(tsd_data, season_weeks = c(20, 21), only_current_season = TRUE)
+  all_onsets <- seasonal_onset(tsd_data, season_weeks = c(20, 21), only_current_season = FALSE)
+
+  expect_equal(current_season, unique(current_onset$season))
+  expect_gt(length(unique(all_onsets$season)), 1)
+})