+++ NEWS.md
@@ -8,6 +8,8 @@
 
 ## Features
 
+* Added a new argument `only_current_season` to `seasonal_onset()`, `seasonal_burden_levels()` and `combined_seasonal_output()` which gives the possibility to either get output from only the current season or for all available seasons (#45).
+
 * Added `combined_seasonal_output()` as the main function to run both `seasonal_onset()` and `seasonal_burden_levels()` to get a combined result for the newest season (#44).
 
 * Added `seasonal_onset()` as a replacement for the deprecated `aedseo()` function (#41).